fix: Solve bug #5
In travelThroughAltitude function in destination.c, set the minimum altitude to INFINITY because this function gets the minimum altitude to travel towards that direction. Change the parameters passed to checkAltitudeInDirection: passed newI and newJ instead of i and j. In advance InDirection function, I passed i and j instead of newI and newJ because we need to change the values of the original altitudes to move. The error occurred when i and j were negatives because it is trying to access a negative position in the array of pointers to the altitudes.